+++ Pitch-Perfect-App/RecordSoundsViewController.swift
@@ -1,0 +1,59 @@
+//
+//  RecordSoundsViewController.swift
+//  Pitch-Perfect-App
+//
+//  Created by Maze Geek on 4/24/18.
+//  Copyright © 2018 Amit Biswas. All rights reserved.
+//
+
+import UIKit
+import AVFoundation
+
+class RecordSoundsViewController: UIViewController {
+    
+    var audioRecorder: AVAudioRecorder!
+
+    @IBOutlet weak var recordingLabel: UILabel!
+    @IBOutlet weak var recordButton: UIButton!
+    @IBOutlet weak var stopRecordingButton: UIButton!
+    
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        stopRecordingButton.isEnabled = false
+    }
+
+
+    @IBAction func recordAudio(_ sender: UIButton) {
+      stopRecordingButton.isEnabled = true
+        recordButton.isEnabled = false
+        recordingLabel.text = "Recording in Progress"
+        
+    let dirPath = NSSearchPathForDirectoriesInDomains(.documentDirectory,.userDomainMask, true)[0] as String
+        let recordingName = "recordedVoice.wav"
+        let pathArray = [dirPath, recordingName]
+        let filePath = URL(string: pathArray.joined(separator: "/"))
+        print(filePath ?? <#default value#>)
+        
+        let session = AVAudioSession.sharedInstance()
+        try! session.setCategory(AVAudioSessionCategoryPlayAndRecord, with:AVAudioSessionCategoryOptions.defaultToSpeaker)
+        
+        try! audioRecorder = AVAudioRecorder(url: filePath!, settings: [:])
+        audioRecorder.isMeteringEnabled = true
+        audioRecorder.prepareToRecord()
+        audioRecorder.record()
+        
+    }
+    
+    @IBAction func stopRecording(_ sender: UIButton) {
+         recordButton.isEnabled = true
+         stopRecordingButton.isEnabled = false
+        recordingLabel.text = "Tap To Record"
+        audioRecorder.stop()
+        let audioSession = AVAudioSession.sharedInstance()
+        try! audioSession.setActive(false)
+    
+  
+    }
+}
+
